Wrapped widget in other parent widget
