+++ src/CMakeLists.txt
@@ -4,7 +4,9 @@ target_sources(megatron3000 PRIVATE
     mainwindow.ui
     histogramwidget.cpp
     histogramwidget.ui
+    transferfunctionrenderer.cpp
     transferfunctionwidget.cpp
+    transferfunctionwidget.ui
     viewport2d.cpp
     viewport3d.cpp
     renderer.cpp
